No media query fallback implementation and removal of @heading mixin
Implementing dual css files (one for mq support, one for when no mq
support is available. Think "below IE9". The class that is to be hooked
in to is defined in _config.scss. Default is ".lt-ie9" which already
exists in the conditional comments. The no-mq CSS file is the existing
"lte-ie8.scss"

The @heading mixin was deemed too confusing to use following real-world
usage of Rock Hammer, and therefore has been removed, with suitable
defaults now residing in _typography.scss for h1-h6. This commit
resolves issues #21 and #29