Do not depend on window load event for responsive nav
responsive-nav.js depends on the window.load which is missed since head.js has been added. Because of this, if you load up rock hammer with a window size small enough that it loads up only the menu link, you'll find that the menu link won't work until you resize the window.

This calls the methods which depend on the window load manually and fixes the issue.